pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jira-mcp"
version = "0.1.0"
description = "MCP server for Jira Cloud REST API v3"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
jira-mcp = "jira_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/jira_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/jira_mcp"]
omit = ["tests/*"]

